fix: move uuid directory inside lib for xcode
