+++ hooks/ios/xcode/lib/pbxProject.js
@@ -19,7 +19,7 @@ var util = require('util'),
     f = util.format,
     EventEmitter = require('events').EventEmitter,
     path = require('path'),
-    uuid = require('./././uuid'),
+    uuid = require('./uuid'),
     fork = require('child_process').fork,
     pbxWriter = require('./pbxWriter'),
     pbxFile = require('./pbxFile'),
