+++ hooks/ios/xcode/lib/uuid/lib/rng-browser.js
