+++ hooks/ios/xcode/lib/uuid/lib/rng.js
