+++ hooks/ios/xcode/lib/uuid/lib/sha1-browser.js
