+++ hooks/ios/xcode/lib/uuid/package.json
