+++ hooks/ios/xcode/lib/uuid/v3.js
