+++ hooks/ios/xcode/lib/uuid/v5.js
